Changes in ContentDialog style.
