+++ src/eShop.UWP/App.xaml
@@ -31,5 +31,6 @@
             </ResourceDictionary.MergedDictionaries>
             
         </ResourceDictionary>
+
     </Application.Resources>
 </Application>
