+++ src/eShop.UWP/Controls/DialogBox.xaml
@@ -1,0 +1,46 @@
+﻿<ContentDialog
+    x:Class="eShop.UWP.DialogBox"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    Background="{ThemeResource SystemControlBackgroundChromeWhiteBrush}"
+    BorderBrush="{ThemeResource SystemControlForegroundAccentBrush}"
+    mc:Ignorable="d">
+
+    <ContentDialog.PrimaryButtonStyle>
+        <Style TargetType="Button">
+            <Setter Property="Padding" Value="0,12" />
+            <Setter Property="Background" Value="{ThemeResource SystemControlBackgroundChromeWhiteBrush}" />
+            <Setter Property="BorderBrush" Value="{ThemeResource SystemControlForegroundAccentBrush}" />
+            <Setter Property="BorderThickness" Value="1" />
+            <Setter Property="FocusVisualPrimaryBrush" Value="{ThemeResource SystemControlForegroundAccentBrush}" />
+            <Setter Property="FocusVisualPrimaryThickness" Value="1" />
+            <Setter Property="FocusVisualMargin" Value="-1" />
+        </Style>
+    </ContentDialog.PrimaryButtonStyle>
+
+    <ContentDialog.SecondaryButtonStyle>
+        <Style TargetType="Button">
+            <Setter Property="Padding" Value="0,12" />
+            <Setter Property="Background" Value="{ThemeResource SystemControlBackgroundChromeWhiteBrush}" />
+            <Setter Property="BorderBrush" Value="{ThemeResource SystemControlForegroundAccentBrush}" />
+            <Setter Property="BorderThickness" Value="1" />
+            <Setter Property="FocusVisualPrimaryBrush" Value="{ThemeResource SystemControlForegroundAccentBrush}" />
+            <Setter Property="FocusVisualPrimaryThickness" Value="1" />
+            <Setter Property="FocusVisualMargin" Value="-1" />
+        </Style>
+    </ContentDialog.SecondaryButtonStyle>
+
+    <ContentDialog.TitleTemplate>
+        <DataTemplate>
+            <Grid>
+                <TextBlock Text="{Binding}" Typography.Capitals="SmallCaps" FontSize="20" FontWeight="SemiBold" TextWrapping="WrapWholeWords" />
+            </Grid>
+        </DataTemplate>
+    </ContentDialog.TitleTemplate>
+
+    <Grid Margin="0,16">
+        <TextBlock Text="{Binding Message}" FontSize="16" TextWrapping="WrapWholeWords" />
+    </Grid>
+</ContentDialog>
